requirements.txt
streamlit>=1.35.0
langchain>=0.2.0
langchain-groq>=0.1.6
langgraph>=0.1.0
langchain-core>=0.2.0
langchain-community>=0.2.0
ddgs>=1.0.0
pypdf>=4.2.0
sentence-transformers>=3.0.0
faiss-cpu>=1.12.0
plotly>=5.22.0
pandas>=2.2.0
python-dotenv>=1.0.1
Pillow>=10.0.0
requests>=2.31.0
